+++ src/sl/saveload.cpp
@@ -611,7 +611,7 @@ void SlSkipBytes(size_t length)
 	return _sl.reader->SkipBytes(length);
 }
 
-int SlReadUint16()
+uint16_t SlReadUint16()
 {
 	_sl.reader->CheckBytes(2);
 	return _sl.reader->RawReadUint16();
@@ -1533,7 +1533,7 @@ void SlSaveLoadRef(void *ptr, VarType conv)
 			break;
 		case SLA_LOAD_CHECK:
 		case SLA_LOAD:
-			*(size_t *)ptr = IsSavegameVersionBefore(SLV_69) ? SlReadUint16() : SlReadUint32();
+			*(size_t *)ptr = IsSavegameVersionBefore(SLV_69) ? (size_t)SlReadUint16() : SlReadUint32();
 			break;
 		case SLA_PTRS:
 			*(void **)ptr = IntToReference(*(size_t *)ptr, (SLRefType)conv);
